+++ src/main/java/standard/StudentManagement/controller/StudentController.java
@@ -1,0 +1,30 @@
+package standard.StudentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import standard.StudentManagement.data.Student;
+import standard.StudentManagement.data.StudentCourse;
+import standard.StudentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.getStudentList();
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentCourse> getStudentCourseList() {
+    return service.getStudentCourseList();
+  }
+}
