+++ src/main/java/standard/StudentManagement/repository/StudentRepository.java
@@ -1,8 +1,11 @@
-package standard.StudentManagement;
+package standard.StudentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import standard.StudentManagement.data.Student;
+import standard.StudentManagement.data.StudentCourse;
+
 
 @Mapper
 public interface StudentRepository {
