+++ src/main/java/standard/StudentManagement/service/StudentService.java
@@ -1,0 +1,32 @@
+package standard.StudentManagement.service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import standard.StudentManagement.data.Student;
+import standard.StudentManagement.data.StudentCourse;
+import standard.StudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> getStudentList() {
+    return repository.search().stream()
+        .filter(student -> student.getAge() >= 30 && student.getAge() <= 39)
+        .collect(Collectors.toList());
+  }
+
+  public List<StudentCourse> getStudentCourseList() {
+    return repository.searchStudentCourses().stream()
+        .filter(studentCourse -> "web開発基礎".equals(studentCourse.getCourseName()))
+        .collect(Collectors.toList());
+  }
+}
